rename the particle_extraction module to particle_region_extraction to match class name property.
